Move trace tests to their own file.
